fix(database): bigint sequence + asyncpg param casts + apply_schema helper
Three fixes surfaced while applying the rebuild to a fresh Supabase project:

1. passages.sequence_number INTEGER → BIGINT. Snapshot values can exceed
   int32 (observed: 1,000,100,010,002 in some hierarchical refs).

2. bootstrap_supabase.py kg_edges INSERT: explicit ::varchar casts on
   $1/$2/$3 to disambiguate asyncpg's parameter type inference (the same
   placeholders are reused on both INSERT values and FK lookup, leading
   to AmbiguousParameterError on the session pooler).

3. apply_schema.py (new): asyncpg-based runner that applies the five
   schema files in order. Drop-in replacement for psql when only Python
   is available, with statement_cache_size=0 for pooler compat.

+++ database/schema/schema.sql
@@ -75,7 +75,7 @@ CREATE TABLE IF NOT EXISTS passages (
     subsection TEXT,
     line_start TEXT,
     line_end TEXT,
-    sequence_number INTEGER NOT NULL CHECK (sequence_number >= 0),
+    sequence_number BIGINT NOT NULL CHECK (sequence_number >= 0),
     text_content TEXT NOT NULL,
     char_length INTEGER CHECK (char_length IS NULL OR char_length >= 0),
     word_count INTEGER CHECK (word_count IS NULL OR word_count >= 0),
